requirements.txt
Django>=2.2
django-background-tasks
intrinio-sdk
